오류사항 수정.

+++ JeJu/src/main/java/com/kgitbank/jeju/dto/TouristSpot.java
@@ -26,6 +26,6 @@ public class TouristSpot {
 	private Date registration_time;
 	
 	private String location;
-	private Float latitude;
-	private Float longitude;
+	private double latitude;
+	private double longitude;
 }
